CLDRex v0.0.1 CLDRex

Provide a list of all supported locales, and a method to check if a given locale is supported.

Types

locale :: atom | String.t

Functions

supported_locale?(locale)

Specs

supported_locale?(locale) :: boolean

Check if a given locale is supported.

supported_locales()

Specs

supported_locales :: list

A list of all supported locales.