CLDRex v0.0.1 CLDRex.Languages

Provide a list of all languages and their localized name.

Types

locale :: atom | String.t

Functions

all()

Specs

all :: Map.t

Return a Map of languages, where each language is localized in its native language.

all_for(locale)

Specs

all_for(locale) :: Map.t

Return a Map of languages localized in the given locale.