CLDRex v0.0.1 CLDRex.Numbers

Provide number localization including conversion to currency, percentage, and decimal values.

Types

locale :: atom | String.t

Functions

localize(number, locale)

Specs

localize(number, locale) :: String.t

Localize the given number into the given locale.

to_currency(number, locale)

Specs

to_currency(number, locale) :: String.t

Convert the given number to the localized currency format.

to_decimal(number, locale)

Specs

to_decimal(number, locale) :: String.t

Convert the given number to the localized decimal format.

to_percent(number, locale)

Specs

to_percent(number, locale) :: String.t

Convert the given number to the localized percentage format.